+++ airslate/exceptions.py
@@ -16,6 +16,7 @@ Classes:
     NotFoundError
     RetryError
     InternalServerError
+    InvalidTypeError
     DomainError
     MissingData
     TypeMismatch
@@ -164,6 +165,15 @@ class InternalServerError(ApiError):
             status=status,
             response=response,
         )
+
+
+class InvalidTypeError(BaseError):
+    """Inappropriate argument type.."""
+    def __init__(self, message=None):
+        if message is None:
+            message = 'Inappropriate argument type'
+
+        super().__init__(message)
 
 
 class DomainError(BaseError):